backend/requirements.txt
fastapi
uvicorn
openai
requests
python-dotenv
pydantic
httpx
